Eliminate the need for RefCell and Rc.
The borrows into the NodePool just needed to be adjusted so that &mut is used
more naturally and so that multiple conflicting &mut borrows are not attempted.

Achieving the latter required not getting borrowing-references to nodes too
soon, to not be holding borrows across unnecessarily longer scopes.  This is now
done by passing their NodeIDs instead of references, to be able to get borrows
of smaller scope only where really needed.  Also, the new NodePool::get_two_mut
method enables getting &mut borrows to two nodes at the same time, instead of
getting one first and then the other next which was causing conflicting borrows.

A few of Node's methods are now moved to be methods of NodePool instead, because
those methods no longer take Node values, because they now take NodeIDs instead
as mentioned above.  This also is more natural since it's a NodePool that owns
all the Nodes and so is what needs to have the &mut borrows into it to get the
borrows to its Nodes.